airflow/dags/requirements.txt
# dags
apache-airflow-providers-http==4.12.0
apache-airflow-providers-postgres==5.11.2
requests==2.31.0
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9 

# dbt
dbt-core==1.8.2
dbt-postgres==1.8.2